+++ src/components/pek2024/Schedule.astro
@@ -1,7 +1,7 @@
 ---
 import ScheduleCard from '~/components/pek2024/ScheduleCard.astro';
 import { type PEK2024Proposal, type PEK2024ProposalList } from '~/types';
-import { getAffiliation } from '~/utils/pek2024/fortee';
+import { getAffiliation } from '~/utils/fortee';
 import axios from 'axios';
 
 type SessionTime = '60min' | '30min';
