+++ src/pages/pek2024/sessions/[id].astro
@@ -6,13 +6,13 @@ import Layout from '../../../layouts/pek2024/PageLayout.astro';
 import { type Tag } from '../../../components/pek2024/SessionInfo.astro';
 import SocialShare from '../../../components/common/SocialShare.astro';
 import { fetchTags } from '../../../utils/fetchTags';
-import { getAffiliation, getJobTitle, getSpeakerProfile, getTalkAbstract } from '../../../utils/pek2024/fortee';
-import { type PEK2024ProposalList } from '../../../types';
+import { getForteeApiUrl, getAffiliation, getJobTitle, getSpeakerProfile, getTalkAbstract } from '../../../utils/fortee';
+import { type PEKProposalList } from '../../../types';
 import { Icon } from 'astro-icon/components';
 
 export async function getStaticPaths() {
-  const { data } = await axios.get<PEK2024ProposalList>(
-    'https://fortee.jp/platform-engineering-kaigi-2024/api/proposals'
+  const { data } = await axios.get<PEKProposalList>(
+    getForteeApiUrl(2024)
   );
   const sessionsList = data.proposals;
   return sessionsList
